+++ lib/src/ping/mac_ping.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:dart_ping/dart_ping.dart';
 import 'package:dart_ping/src/response_parser.dart';
@@ -32,10 +31,10 @@ class PingMac extends BasePing implements Ping {
       summaryStr: _summaryStr);
 
   @override
-  Future<Process> get platformProcess async {
+  List<String> get params {
     var params = ['-n', '-W ${timeout * 1000}', '-i $interval', '-m $ttl'];
     if (count != null) params.add('-c $count');
-    return await Process.start(ipv6 ? 'ping6' : 'ping', [...params, host]);
+    return params;
   }
 
   @override